Add initial builders.
Most of these are variants of "dotnet msbuild"; most arguments aren't hooked up yet.
"dotnet restore" may also be relevant to have available.

It may also be better to have the display names all start with ".NET: " so that they group together in the drop-down of the job configure UI.

Best might be to have a single "dotnet" builder, providing a superset of their command line options, but more jelly-fu is needed to make the UI for that work nicely.

There is also a side issue with the fact that running "dotnet.exe" will, on Windows at least, spawn a bunch of background processes (with no means of preventing that). Those would then make that first job "hang"; further jobs would be fine (reusing the processes previously spawned).