Implement the actual execution.
It uses a script (based on the general behaviour of the existing bat/sh builders).

Made the 'sdk' field optional (which means pipelines can just use 'dotnetBuild', possibly adding 'dotnetsdk "My SDK"' to the tools section).

Dropped fields from the 'build' builder.
Additional parameters for all builders to be wired in later, once the UI is done for all of them.