Implemented a tool installer.
This uses a JSON file containing the available SDKs installation
packages. Currently this is a hand-made file; I did't find an official
file from Microsoft. It's also incomplete (it gets tedious; a dozen
releases, each with a handful of feature band SDKs, for which there are
8 or so installation packages).

The existing DowloadFromUrlInstaller is sadly not applicable - it is
specifically intended for cases where the download package is not
platform-specific. So this installer mixes bits of that with the more
generic ZipExtractionInstaller and adds a convenient UI for selecting
the installation package to use.