Rework the builders.
These now create and run actual commands, using ArgumentListBuilder,
instead of creating a batch/or shell script.
This works even when a wrapper is used, because the home environment
variable it sets is picked up and used to construct the full command
path.

Only major remaining problem is that when run from a pipeline,
environment set by a wrapper is NOT available, so it will run 'dotnet',
which then runs the system-wide one (again, because the updated PATH is
not picked up). Apparently this requires implementing Step separately,
which seems to make SimpleBuildStep useless (and kind of dangerous,
since it essentially ignores wrappers).

Other associated tweaks:
- moved writing of the command completion message to DiagnosticScanner,
  including ownership of an encoded CompletionNote
  - this uses a specific marker to indicate where in the message the
    exit code starts; slightly nicer for l12n, but still depends on it
    being at the end of the message
- added an encodeNote() method to DotNetUtils to avoid code duplication
  - obsoletes DiagnosticNote::createEncoded, which has been removed
- added DotNetUtils::RESOLVE_NOTHING, a no-op VariableResolver<String>
- added log message when removing global.json (helps verify that this
  happens correctly at the end of a wrapper's scope)
- corrected the default for the specificSdkVersion setting in the
  builders (it was on by default)