Fix DotNetStep.
The snippet generator wasn't working, because the builder field was not
public.

Also, I could have sworn it did not show up in the snippet generator as
its own separate entry, but it does, so I added a help file to specify
that it's not intended for direct use (not that you can do anything with
it other than making a call to dotnetBuild & co look more complicated).