Update the JSON files.
They now use WIP output of the crawler written for the update center;
this includes changes to the format for the RID catalog and the SDK
downloads.
Next step, once that crawler is approved, is to switch the code to use
the downloadables instead of resources.

Dropped the JSON schema files.