Switch to CustomDescribableModel
The deprecation of the properties works for the lists/arrays, but not
for the maps. So instead, CommandDescriptor now implements
CustomDescribableModel (using the default implementations) so that
specific descriptors can override customUninstantiate to filter out
"duplicate" properties, leaving only the one to be suggested for
pipeline use by the snipper generator.