Don't opt-out of telemetry at SDK level by default
Now that there is a global config setting that can opt out for all SDKs
at once, there is no need to default the opt-out for each SDK separately
anymore. In fact, the actual object property already defaults to `false`
now, but the checkbox in `config.jelly` was still defaulting to checked.