Add a default entry to the wrapper's SDK selection
Now there will always be a "(Not Selected)" entry at the top of the
list. This helps with the form validation; before, if only one SDK was
configured, the validation would always show the "Required" error,
because the property was blank, and the selection of the single list
item didn't take effect until either another entry (which wasn't there)
was selected or the form was submitted.