JENKINS-68729: Abort the build on error
Previously, when there were errors (either reported by a build, or a
non-zero return code), the build was marked as failed, but continued to
the next step.

Now, such errors will cause the build to be aborted.

A new "Continue On Error" option has been added to provide the old
behaviour. However, it was _not_ made the default; this may break
compatibility with any existing builds that rely on the old behaviour.