Disable the existing build steps
This marks all existing builders as not applicable to any projects,
effectively making them unavailable from the Jenkins UI. A deprecation
banner was added to the configuration UI as well.

Dropped the "should this command be allowed" configuration - none of
them are allowed now. That leaves only the global telemetry opt-out.

Dropped the `configuration` field from `DotNetSDK.DescriptorImpl`; it
caused the configuration to be serialized there. The telemetry opt-out
check now retrieves the configuration at run time.

Note: there seems to be no way to mark the pipeline steps as advanced or
deprecated (to move them "out of the way" in the snippet generator).