Move `DotNetArguments`
This class now lives in the top-level package. To preserve
backward compatibility, the version in the old location is now a
deprecated wrapper around the new one.

This will make it easier to drop the old commands at a later point,
because everything in the commands package will be safe to remove.